Further refactor of packages
