+++ manipulation/src/functTest/java/org/jboss/gm/manipulation/MultiModuleFunctionalTest.java
@@ -12,8 +12,8 @@ import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 import org.gradle.testkit.runner.TaskOutcome;
+import org.jboss.gm.common.io.ManipulationIO;
 import org.jboss.gm.common.model.ManipulationModel;
-import org.jboss.gm.common.utils.ManipulationUtils;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -28,7 +28,7 @@ public class MultiModuleFunctionalTest {
         final File multiModuleRoot = tempDir.newFolder("multi-module");
         TestUtils.copyDirectory("multi-module", multiModuleRoot);
         assertThat(multiModuleRoot.toPath().resolve("build.gradle")).exists();
-        final ManipulationModel alignment = ManipulationUtils.getManipulationModel(multiModuleRoot);
+        final ManipulationModel alignment = ManipulationIO.readManipulationModel(multiModuleRoot);
 
         final BuildResult buildResult = GradleRunner.create()
                 .withProjectDir(multiModuleRoot)
