+++ manipulation/src/functTest/java/org/jboss/gm/manipulation/SimpleProjectFunctionalTest.java
@@ -12,8 +12,8 @@ import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 import org.gradle.testkit.runner.TaskOutcome;
+import org.jboss.gm.common.io.ManipulationIO;
 import org.jboss.gm.common.model.ManipulationModel;
-import org.jboss.gm.common.utils.ManipulationUtils;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -29,7 +29,7 @@ public class SimpleProjectFunctionalTest {
         TestUtils.copyDirectory("simple-project", simpleProjectRoot);
         assertThat(simpleProjectRoot.toPath().resolve("build.gradle")).exists();
 
-        final ManipulationModel alignment = ManipulationUtils.getManipulationModel(simpleProjectRoot);
+        final ManipulationModel alignment = ManipulationIO.readManipulationModel(simpleProjectRoot);
 
         final BuildResult buildResult = GradleRunner.create()
                 .withProjectDir(simpleProjectRoot)
