+++ manipulation/src/main/java/org/jboss/gm/manipulation/ManipulationPlugin.java
@@ -12,9 +12,9 @@ import org.gradle.api.Project;
 import org.gradle.api.plugins.MavenPlugin;
 import org.gradle.api.publish.maven.plugins.MavenPublishPlugin;
 import org.jboss.gm.common.Configuration;
+import org.jboss.gm.common.io.ManipulationIO;
 import org.jboss.gm.common.model.ManipulationModel;
 import org.jboss.gm.common.utils.ManifestUtils;
-import org.jboss.gm.common.utils.ManipulationUtils;
 import org.jboss.gm.manipulation.actions.ManifestUpdateAction;
 import org.jboss.gm.manipulation.actions.MavenPublicationRepositoryAction;
 import org.jboss.gm.manipulation.actions.OverrideDependenciesAction;
@@ -41,7 +41,7 @@ public class ManipulationPlugin implements Plugin<Project> {
     @Override
     public void apply(Project project) {
         // get the previously performed alignment
-        final ManipulationModel alignmentModel = ManipulationUtils.getManipulationModel(project.getRootDir());
+        final ManipulationModel alignmentModel = ManipulationIO.readManipulationModel(project.getRootDir());
         final ManipulationModel correspondingModule = alignmentModel.findCorrespondingChild(project.getName());
 
         final ResolvedDependenciesRepository resolvedDependenciesRepository = new ResolvedDependenciesRepository();
